cec-client: call stop on the webserver when ctrl-c is caught
There was an issue in the signal trapping where the app didn't stop. I
suspect it was either that the signal was caught by the asio library, or
simply that the server wasn't stopped.

Stop the webserver when ctrl-c is caught in an attempt to keep zombie
processes around or the program from not quitting.