cec-client: fix invalid retries after a failed send
Sending a single response worked. But the m_curl settings weren't reset
after a send. This meant that the next time a "GET" was performed, it
would also perform a "PUT" and everything would get confused.

Properly clean up the m_curl interface after each time it gets invoked
so that it... well... works.